chore(stack): validate path and root_directory against allowed characters (#1956)
This PR adds limits to the length of input paths corresponding to the limits of docker containers and EFS volumes. It also ensures that paths do not include special characters beyond a small allowlist of common characters and the `/` separator. 

<!-- Provide summary of changes -->

<!-- Issue number, if available. E.g. "Fixes #31", "Addresses #42, 77" -->

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.