chore(template): fix statemachine and sidecar so e2e tests pass (#1977)
The statemachine's `DefinitionSubstitutions` sections expects strings as values. In #1968, we updated them to be a list  and broke the template.  
We also added `"-}}"` in our service templates above the sidecar template which removed extra whitespaces resulting in the template for sidecars to be malformed, this change adds the new lines back.



By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.