fix(cicd): truncate pipeline name to prevent failures (#1957)
When building pipelines, we are limited by character maximums for multiple strings:
128 for PipelineRolePolicy, BuildProjectPolicy, etc., and the CODEBUILD_BUILD_ID tag for Docker.
100 for CodePipeline pipeline names.

So that pipelines built from repos with very long names and/or repo provider accounts with very long names do not fail, this change reduces the characters when we generate pipeline names, and throws an error when pipeline names exceed 100 characters in `pipeline update` (in case pipeline names were edited in the pipeline manifest). We limit to 100 to satisfy CodePipeline's pipeline name limit (100), but also to enable the automated tacking on of `-CodepipelinePolicy` and `-BuildprojectPolicy` (and any future policy name suffixes) to the end of the stack names (128). 

The extra characters are omitted _from the front_; at the moment, pipeline names go from general ("pipeline") to specific (repoName or even branchName, if users are [changing pipeline names to enable multiple pipelines from a single repo](https://github.com/aws/copilot-cli/discussions/1925)) so truncating from the front of the string preserves the more valuable information.

In the buildspec, we truncate the string yet again, this time to 128 characters. Again, we cut off excess characters from the front. Here, it is because `-BuildProject-[UUID]` is appended to form the CODEBUILD_BUILD_ID, and we don't want to lose the identifying UUID. We must truncate in two steps because it is possible for a pipeline name (of, say, 100 characters), to successfully generate a CFN stack, but fail in CodeBuild because with the added characters it becomes 150 characters long. 

Fixes #1939.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.