feat(manifest): enable "network" field to specify subnets and sec groups (#1979)
The manifests now support the following new fields:
```yaml
network:
   vpc:
      placement: 'public|private'  # which subnets to place the tasks.
      security_groups: [string]    # list of additional security groups.
```

The `network.vpc.placement` field won't be really useful atm for customers that use Copilot generated VPCs since we don't create NAT gateways. However, it will be supported in the future. This is milestone 1. of #1959.  
On the other hand, customers that imported their VPC while creating an environment will now be able to place their tasks in private subnets.

The security groups feature is available for everyone to attach additional security groups to their tasks.

Related #1959
Closes #1453

_By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice._