chore: state machine should launch with PV 1.4.0.  (#1993)
<!-- Provide summary of changes -->
In order to enable EFS file systems, we need to launch scheduled jobs from the state machine using PV 1.4.0. 

<!-- Issue number, if available. E.g. "Fixes #31", "Addresses #42, 77" -->
Related #1559 

I tested this in a personal account and confirmed that a scheduled job could mount, modify, and list out a file system when `read_only` was set to `false`.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.