chore(docs): update docs and prompt to emphasize pushing of necessary files (#1981)
Previously, only the buildspec.yml needed to be added to the source repo in order to build a pipeline. But now the buildspec checks the pipeline manifest to see which environment stages to deploy to (and only upgrades those envs), so both files (and `.workspace`)need to be checked into the user's repo. This updates docs and CLI prompts to reflect that change.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.