chore(manifest): parse strings using shell-style rules (#2014)
In #1998, the way that the strings are split into string slice may cause bugs when passing the values to cfn template. This PR implements a similar fix as #1467.

Fixes #1998 

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.