chore(cli): env init add env to stackset before env stack creation (#2010)
<!-- Provide summary of changes -->
Currently we create the env stack before adding the env to the app stack set, which means the ECR repo will be created after the env stack creation. This becomes a problem when we want to upload the env stack custom resources script to the s3 bucket created by the app stack set.

With this PR, we swap the order so that the env will be added to the stackset before creating the env stack, making it possible to upload the custom resource scripts to the app s3 bucket.
<!-- Issue number, if available. E.g. "Fixes #31", "Addresses #42, 77" -->

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.