chore(cli): display public ip address for `task run` (#2032)
This PR displays public IP addresses of tasks after `task run` succeeds.

An allow-action `ec2:DescirbeNetworkInterfaces` is added to environment-manager-role in support of this functionality in the case where tasks run in an environment.

If we are not able to retrieve the IP addresses for some reasons (e.g. the ENI information is not found for the task, or the public ip associated with the ENI is not found), we will just not show the IP address instead of erroring it out explicitly. For more information see the discussion below.

Screenshot of cli output:
`copilot task run --app project --env env-with-describe-eni` which runs a task in an environment with the updated `DescribeNetworkInterfaces` permission
<img width="665" alt="Screen Shot 2021-03-10 at 1 55 55 PM" src="https://user-images.githubusercontent.com/79273084/110689758-b076da00-81a8-11eb-8ac2-fefaff1bf071.png">

`copilot task run --app project --env env-with-describe-eni --count 3`  which runs 3 tasks in an environment with the updated `DescribeNetworkInterfaces` permission
<img width="614" alt="Screen Shot 2021-03-10 at 1 59 28 PM" src="https://user-images.githubusercontent.com/79273084/110690009-0481be80-81a9-11eb-9417-d33b1394ed2f.png">

`copilot task run --app project --env env-with-o-describe-eni`  which runs a task in an environment created *before this pr's changes* (no `DescribeNetworkInterfaces` permission)
<img width="544" alt="Screen Shot 2021-03-10 at 1 42 09 PM" src="https://user-images.githubusercontent.com/79273084/110690182-3dba2e80-81a9-11eb-8f2c-0c47cf52da13.png">


Addresses #1269 

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.