ignore INACTIVE cluster in DefaultCluster
When we delete ECS cluster, AWS return it as INACTIVE cluster.

```
INACTIVE
The cluster has been deleted. Clusters with an INACTIVE status may remain discoverable in your account for a period of time. However, this behavior is subject to change in the future, so you should not rely on INACTIVE clusters persisting.
```
https://docs.aws.amazon.com/AmazonECS/latest/developerguide/clusters.html

We can't run task on INACTIVE cluster so we should ignore it.