chore: specify which targets are PHONY (#2049)
Previously, we had several targets that were just recipes names and did
not produce any artifacts. Adding `.PHONY` ensures that these are names
and not file targets, thus guarding against the albeit
unlikely chance that there is a file added with the same name as the
recipe.

https://www.gnu.org/software/make/manual/make.html#Phony-Targets

<!-- Provide summary of changes -->

<!-- Issue number, if available. E.g. "Fixes #31", "Addresses #42, 77" -->

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.