feat(cli): allow to reject any ssm plugin install/update (#2060)
<!-- Provide summary of changes -->
This PR makes it possible to reject any SSM plugin install/update with `--yes=false`. Technically not many users want to use this feature. It is mainly for the e2e tests so that we can skip without updating the SSM plugin.
<!-- Issue number, if available. E.g. "Fixes #31", "Addresses #42, 77" -->

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.