test(e2e): fix exec suite by disabling exec log test (#2136)
<!-- Provide summary of changes -->
Disable the exec log test as the log stream name created by ECS is not stable (a bug to be fixed by ECS). We'll enable it again once the issue is fixed by ECS.
<!-- Issue number, if available. E.g. "Fixes #31", "Addresses #42, 77" -->

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.