chore: render only env controller action if env stack has no updates
Previously, the progress tracker assumed that if the EnvControllerAction
is triggered with new Parameters, then the EnvStack gets updated.
However, that's not necessarily true. The EnvControllerAction can be
triggered yet not update the EnvStack resources.

In this change, we're introducing a new component
`envControllerComponent` that renders the env stack changes if there are
any, otherwise defaults to rendering the env controller action resource.