chore: render only env controller action resource if the env stack has no updates  (#2150)
Previously, the progress tracker assumed that if the EnvControllerAction is triggered with new Parameters, then the EnvStack gets updated.  
However, that's not necessarily true. The EnvControllerAction can be triggered yet not update the EnvStack resources.

In this change, we're introducing a new component `envControllerComponent` that renders the env stack changes if there are any, otherwise defaults to rendering the env controller action resource.

_By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice._