fix(cli): prevent pipeline status errors (#2170)
This fixes two bugs in `pipeline status`-- previously, the command was throwing a nil pointer error when the prompter was being used, and `describe` threw an error when there was no pipeline name. It was also returning the pipeline status of the pipeline found in the manifest if there were no pipelines associated with the app name passed in with the flag. Now the command behaves as expected.

Fixes https://github.com/aws/copilot-cli/issues/2004.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.