docs: add new count and range subfields to manifest docs (#2206)
Closes https://github.com/aws/copilot-cli/issues/2205

NOTE: This change also includes a refactor that makes use of the `mkdocs-macros` pluging. This allows embedding markdown files into other ones, which allows us to de-dupe the docs for the two types of service manifests.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.