fix: return service config if no env overrides specified (#2227)
Previously, if an env name was specified in a service manifest with no
subfields, `svc deploy` would panic. This checks if the override config
for an environment is a nil pointer and returns early if it is.

Fixes #2226


By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.