test(e2e): remove FS size check due to eventual consistency problems (#2313)
<!-- Provide summary of changes -->
This PR cleans up an e2e test case which relied on EFS updating the metered size of a newly created file system. This led to timeouts and eventual consistency issues. This check was overkill, since we already verify that the EFS filesystem is created and visible in the output of Env Show.

```
2021-05-13T09:30:27.982-07:00 | EFS filesystem size should be greater than 1 MiB [It]
  | 2021-05-13T09:30:27.982-07:00 | /github.com/aws/copilot-cli/e2e/multi-svc-app/multi_svc_app_test.go:356
  | 2021-05-13T09:30:27.982-07:00 |  
  | 2021-05-13T09:30:27.982-07:00 | Timed out after 240.000s.
  | 2021-05-13T09:30:27.982-07:00 | Expected
  | 2021-05-13T09:30:27.982-07:00 | <bool>: false
  | 2021-05-13T09:30:27.982-07:00 | to be true
```

<!-- Issue number, if available. E.g. "Fixes #31", "Addresses #42, 77" -->

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.