chore: add Job functions to deploy package (#2276)
<!-- Provide summary of changes -->
This is a prerequisite PR for the ongoing `job logs` work. This adds two methods to `deploy` to allow Copilot to check whether a given job is deployed in an environment, and to list all deployed jobs by app and env. 

This PR needs to be merged before I can finish work on `job logs Ask`.

Related #2272 

<!-- Issue number, if available. E.g. "Fixes #31", "Addresses #42, 77" -->

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.