fix: add default mapping and integ tests for addons (#2346)
<!-- Provide summary of changes -->
This PR adds a default mapping for the DBMaxCapacity and DBMinCapacity values which doesn't depend on customers having run `env init` before `storage init`. The default value `All` can't collide with an environment name because it contains a capital letter and this is disallowed at env creation. 

I've also added integ tests to check the outputs of the addons `BinaryMarshaler`s. 

<!-- Issue number, if available. E.g. "Fixes #31", "Addresses #42, 77" -->

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.