feat(network): allow imported VPCs with zero public subnets (#2356)
Previously, if users imported their own VPC resources, Copilot assumed they were bringing exactly two public and two private subnets and would error out if this was not the case.

These changes allow for more flexibility, with the warning that load-balanced web services cannot be deployed in an environment with fewer than two public subnets, as the Load Balancer requires two in different AZs. 
I believe we do not need to warn users that it's impossible to deploy workloads with 'private' network placement specified if their VPC does not have private subnets, and the same with 'public.'

I have manually tested:
- importing a console-generated VPC with 0 subnets
- importing a console-generated VPC with 1 public subnet
- importing a console-generated VPC with 1 private subnet
- importing a Copilot-generated VPC with: 1 public, 2 public, 1 private, 2 private, 1 public + 1 private
- deploying a load-balanced web service (public placement) to the five configs above
- deploying a backend service (public placement) to the five configs above
- deploying a load-balanced web service (private placement) to the five configs above
- deploying a backend service (private placement) to the five configs above

To-do:
- [x] Make the same changes for jobs
- [x] Make sure `env update` works without public subnets
- [x] Make sure these changes are backwards-compatible with previously imported VPCs

To-do in a future PR: 
- Create a docs page for custom envs (https://github.com/aws/copilot-cli/issues/2224)

Addresses #2079, #2233.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.