test(e2e): add skeleton for pipeline e2e test (#2422)
This initial e2e/pipeline pkg just creates a codecommit repository
where we can push the service code. Once this PR is merged and we know
that it works with CodeBuild, we'll send a follow-up PR to actually run
the "copilot pipeline" commands.

- [x] Tested the changes against my AWS account and local machine

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.