fix(manifest): preserve default network config on empty env overrides
Fixes #2420

Previously, when a user provided a network config and an env override:
```yaml
network:
  vpc:
    security_groups: ['sg-123123']
environments:
  dev:
    count: 2
```
The network config for the "dev" environment got overriden to an empty
struct{}.

This change ensures that the default entry is preserved for network
config is preserved for the "dev" environment.
This is done by changing the type for the Network field to a pointer so
that mergo does not overwrite it.