chore: detect os/arch and warn if ARM arch (#2545)
This change logs a warning and suggests a workaround for ARM machines. 

Tested that the warning was logged on ARM arch, and also that `svc deploy` worked with the workaround.
Next step: target platform amd64/x86 for ARM builds.

(This includes an unrelated addition to the Makefile.)

Related: #1949, #2369.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.