chore(stack): error if target sidecar container has no port (#2571)
<!-- Provide summary of changes -->
After #2565, it is possible that a sidecar container has no port open. However, we should error out early if that sidecar container is the target container for LB.
<!-- Issue number, if available. E.g. "Fixes #31", "Addresses #42, 77" -->

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.