fix(lambda): handle gracefully for same alias default domain (#2607)
<!-- Provide summary of changes -->
fixes #2602. Currently Copilot fails if users set the `alias` to be the same as the default domain name we assign them. Though this is not a very common behavior, we should handle this edge case gracefully. 
<!-- Issue number, if available. E.g. "Fixes #31", "Addresses #42, 77" -->

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.