feat: build ARM-architecture-based images as AMD64 (#2636)
Previously, users could not build ARM-based images with Copilot (without employing workarounds) because Fargate does not currently support ARM-based tasks. This PR automatically builds images as `linux/amd64` when ARM (or other non-AMD) architectures are detected, using Docker's `--platform` flag. This is a best-effort redirect, as users' Docker servers must be multi-platform-capable in order for `--platform` to work. A warning is logged at `svc init` and info is logged at `svc deploy` IF non-amd64 architectures are detected.

An additional manifest field, `platform:` is surfaced iff OS/arch is being redirected, to provide visibility to the user and for extensibility (we can take changes in the future). Validation is performed when the manifest is unmarshaled. The field currently takes a string, but will accept a map in a future PR. 

_The workflow/experience does not change for amd64 users!_

Related: #2545; closes: #1949, #2369.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.