chore: fix svc show should list proper endpoint under "service discovery" section (#2676)
<!-- Provide summary of changes -->
UI bug. svc show should display the env var as well as the "Service Discovery" section correctly. 

This PR adds a call to the env describer under the hood of the service describer clients to get the correct endpoint for the environment, service, and port. 

<!-- Issue number, if available. E.g. "Fixes #31", "Addresses #42, 77" -->

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.

<img width="970" alt="Screen Shot 2021-07-28 at 8 57 06 PM" src="https://user-images.githubusercontent.com/25392995/127429366-73caaad1-24d3-4311-9d88-3e1e7a33d1e0.png">