chore(manifest): refactor pointer struct to struct in manifest (#2798)
This PR is a refactor that removes unnecessary pointers ™️ from manifest.

"Necessary pointers" ™️  to the following structs are not removed:
##### <s>`ContainerHealthCheck`</s> and `PlatfromArgsOrString` for all workloads (e.g. https://github.com/aws/copilot-cli/blob/ff0a32e3d0d6821408ba6f05fcad6a0450f75f24/internal/pkg/manifest/lb_web_svc.go#L64)
This is because of the way we marshal manifest object into YAML during `svc init` or `job init`.

<s>##### `EFSConfigOrBool`
A manifest input that has:
```
efs:
<the rest of the manifest>
```
is deemed invalid, and we error out during the conversion from a manifest object to a template object. We cannot remove the pointer right now because otherwise we cannot differentiate a said invalid input, and a valid input that didn't specify `efs` at all. 

This validation should instead happens at an earlier stage, potentially from within a dedicated `manifest.validation` package. We can remove the pointer after we can move the validation to the said earlier stage.</s>

##### `SQSQueue `
For this particular manifest field, a manifest input that has:
```
topics:
     - name: topic1
        queue:
<the rest of the manifest>
```
is deemed valid. If the `queue` field is specified without any value, we create the queue for that topic with default configurations. Hence, there is indeed a need to differentiate an empty queue from an absent queue. 

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.