fix(manifest): remove unnecessary `allowed_workers` field (#2821)
There is no way today to limit an SNS topic subscription to be based on the tag of SQS queues (https://aws.amazon.com/premiumsupport/knowledge-center/sqs-sns-subscribe-cloudformation/). Instead, by default SNS will limit subscriptions to be only within the same AWS account.  

Therefore, the `allowed_workers` field is not needed. This is not a breaking change, it makes a required field now obsolete. 


_By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License._