chore: fix integration test by unsetting `AWS_DEFAULT_REGION` (#2845)
The previous attempt #2837 to fix the unit test failed because `AWS_SESSION_TOKEN` is not the true culprit of the failure. As described in this [doc](https://docs.aws.amazon.com/codebuild/latest/userguide/build-env-ref-env-vars.html), `AWS_DEFAULT_REGION` is set in CodeBuild environment, which should be the reason why the test is failing.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.