chore: friendly message when alias is used while domain is not associated with the app (#2911)
We should display a friendly message when we detect that `alias` is specified in manifest while the app is not associated with a domain.

Previously, when this happens, an RDWS deployment fails when CloudFormation tries to create the custom resource. On the other hand, for LBWS, we swallow the alias input when domain is not associated. 

Therefore in this PR, we error out early for RDWS while displaying the message. For LBWS, we only display messages while not failing.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.