chore: return subnet CIDR block information from `ec2` clients (#2945)
This PR is to support the NLB feature. When creating NLB, we need to allow traffic from the subnets where the NLB lives in into the service. To create a security group that allows such an ingress, we need to provide the CIDR block information of the subnets.

Preceding: #2916 

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.