chore: fix e2e test by fixing s3 bucket template (#2975)
The previous PR #2974 causes e2e test to fail because it only changes s3 bucket's logical ID, but didn't update the references to the resource. This PR fixes that. 

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.