chore: fix e2e test by skipping `WorkspaceNotFound` error in `svc init` and `job init` (#2982)
The e2e test failed because of #2966. When `copilot` commands is run for the first time in a directory, the workspace, indicated by the existence of `copilot/` folder,  may not exist. This error should be skipped, not surfaced, as well.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.