chore: quick follow-up to revert partial changes to buildspec (#2994)
This PR is a quick follow up to #2970, reverting some changes. 

The previous PR assigns different tags to each environment image. However, if the images are the same across all environments, the tags shouldn't be different.

We will send out a follow-up PR to update the `buildspec` such that we assign the same tag / environment-specific tags to images depending on whether the images are the same across environments / environment-specific.

For now, we temporarily revert the change so that it maintains existing behavior until we send out an update.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.