chore(pipeline): exit build stage if manifest validation failed (#3080)
Fixes #3072.

If a workload manifest fails, the pipeline build stage will fail, too, but only after many more steps, logs, and time. The validation error gets buried, and it isn't obvious that it's the cause of the failure.

This change checks if the `params.json` file that would be generated upon successful manifest validation is empty; if it is, the build exits with a suggestion to look for a manifest validation error message. There are, of course, still logs to wade through, but this should make the point of failure easier to pinpoint.

Manually tested with invalid service manifest (ARM + Spot), valid service manifest, invalid job manifest (bad platform), valid job manifest.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.