chore: recommended action when domain registered with the app is not in Route53 (#3081)
Nowadays people can associate a domain with their application with `copilot app init --domain example.com`. At this moment,  it is required that the domain is registered with Route53 in order for future operation to work (e.g. enabling HTTPS). However, this information is not well surfaced right now.

This PR provides a better recommended action when it is detected that the domain isn't registered with Route53.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.