chore: replace error with a friendly warning if domain is not explicitly owned (#3090)
Instead of error-ing out, Copilot displays a friendly warning with purposed solution if the domain is not explicitly owned.

Sometimes the user put a subdomain for `--domain` flag, and they does own the top-level domain. In this case, the domain may not be explicitly owned by the account, but things should work as long as there is an appropriate NS record delegating the subdomain to the current account.

When the domain is indeed registered outside of Route53, they do need to transfer the domain to Route53 - otherwise later operations would require some manual intervention for the domain to work.

Either way, it's okay if they fix the problem after the application is created and registered with the domain. This way they don't have to tear down the app, fix the problem, and create the app again.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.