chore(custom-resources): retrieve validation options that need to be deleted (#3108)\n\nThis PR:
1. Add skeleton for "Delete" action that deletes unused certificate and records 
2. Implement functions to determine which validation options need to be deleted

### What resources are being managed
The custom resource manages the certificate, the validation options, and the aliases. Upon "Delete" event, it needs to delete:
1. The certificate that is used by the service with the specified set of aliases;
2. The validation records that will be unused after deleting the certificate;
3. The A records.

Note that 1.[The certificate that is used by the service with the specified set of aliases] is determined by:
1. The canonical certificate domain name. Copilot only creates certificate using the name "svc-nlb.env.app.domain", so it also only deletes certificate with such name.
2. The tags. Copilot should only be able to delete certificate that's Copilot-tagged.
3. The SANs. If a "Delete" event is sent with a list of alias, that means the service doesn't need this specific combination of aliases anymore. The certificate used to validate this combination is no longer needed.

A certificate satisfying all three of these conditions should be safely deleted.

### Validation Options
After deleting the certificate, the validation options used to validate the certificate need to be cleaned up. However, we should make sure that the option being deleted is not used by other certificate, other service.

A validation option can be deleted only if:
1. The option's domain is one of the three Copilot recognized pattern: a.env.app.domain, a.app.domain, or a.domain. The hosted zones associated with these three patterns are Copilot-managed; by ensuring the domain follows one of these patterns, Copilot is only touching the records in the authorized hosted zones. 
2. Among all Copilot-tagged service certificate, it is used to validate only the certificate that is pending deletion. 
3. It is not currently in use by any other service.

Note that 1. and 2. are virtually the same requirement as indicated by the current ALB implementation.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.
\n