chore: push template to s3 and deploy workload stack (#3183)
This PR is to address [this comment](https://github.com/aws/copilot-cli/pull/3171#discussion_r781820226) in a previous PR. When deploying workloads, instead of using the template body string, we first upload the template to the s3 bucket, and then refer to the URL in the API call. This is similar to the current approach to deploy a pipeline.

We need to do this because there is a 51KB limit of template body string, but 1MB if it's uploaded to s3 bucket. As the template is growing larger and larger, 51KB may become insufficient.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.