chore: patches to NLB support  (#3187)
This PR contains a bunch "patch fixes" (quoted because they are not really fixes since NLB isn't released) that surfaced after a few test runs. Each commit fixes an issue, and the commit message should be a summary of what it's fixing.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.