feat(pipelines): Added partial support for arm64 platform in pipelines (#3190)
# Summary of changes:

This PR adds support for ARM64 based builds in the **pipeline** command by adding a new **EnvironmentType** which is automatically selected if the image name contains the string **aarch64**

It also updates the **docker runtime-version** to **19** as **18** isn't supported in ARM_CONTAINER at this time.

One thing this patch doesn't solve is to create a different **buildspec.yml** to point to the arm64 binaries for the copilot tool. This is because the lifecycle for the files seems to me to be different and I wasn't sure what is the best way to address this. At the moment, you need to manually change the following lines in your project ```buildspec.yml```:

``` yaml
phases:
  install:
    commands:
...
      - wget https://ecs-cli-v2-release.s3.amazonaws.com/copilot-linux-v1.13.0
      - mv ./copilot-linux-v1.13.0 ./copilot-linux
```

to

``` yaml
phases:
  install:
    commands:
...
      - wget https://ecs-cli-v2-release.s3.amazonaws.com/copilot-linux-arm64-v1.13.0
      - mv ./copilot-linux-arm64-v1.13.0 ./copilot-linux
``` 

This PR addresses #3158

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.