Revert "feat(pipelines): add branch detection to `pipeline init`" (#3203)
Reverts aws/copilot-cli#3182.

Need to address the fact that the value expected to be passed in with the `--url` flag is of the format of URLs passed in with `git clone...` and output by `git remote -v`, NOT what users may consider their repo URL, the browser URL. For branch detection, we need to identify that the indicated repo does exist as a local git remote.

Possible solutions are parsing the flag input to widen the validation, allowing the remote name (short name) to get passed in instead, etc.