feat(cli): add `--container-insights` flag to `env init` (#3126)
I've added support for Container Insight for the next version of Copilot :) 
You can override the value via the CLI if you don't want to activate it, but it will be activated by default. 

#1579 

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.