chore: use conditional before `!Split` to pass empty list (#3214)
This is a bug found from a recent manual test. 

`!Split` returns an array of one empty string if given an empty string. This is not we want - if given an empty string, we want `Aliases` to receive an empty array.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.